Extract test helpers from 'FeedUIIntegrationTest' to their own files
